shifaa-ia/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0.post1
pydantic>=2.4.2
pydantic-settings>=2.0.3
google-genai>=1.0.0
openai>=1.3.5
python-dotenv>=1.0.0
